# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DEX Studio",
    "description": "DataEngineX control plane",
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/scheduler/trigger/{pipeline}": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "Manually trigger a pipeline run",
        "operationId": "api_scheduler_trigger_api_scheduler_trigger__pipeline__post",
        "parameters": [
          {
            "name": "pipeline",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pipeline"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Api Scheduler Trigger Api Scheduler Trigger  Pipeline  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/pipelines/{name}/run": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "Ad-hoc pipeline run",
        "operationId": "api_pipeline_run_api_pipelines__name__run_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Api Pipeline Run Api Pipelines  Name  Run Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pipelines/{name}/backfill": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "Trigger pipeline backfill",
        "operationId": "api_pipeline_backfill_api_pipelines__name__backfill_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Api Pipeline Backfill Api Pipelines  Name  Backfill Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/watermarks/{source}": {
      "delete": {
        "tags": [
          "api"
        ],
        "summary": "Reset watermark (triggers reprocessing)",
        "operationId": "api_watermark_reset_api_watermarks__source__delete",
        "parameters": [
          {
            "name": "source",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Api Watermark Reset Api Watermarks  Source  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/schema/{pipeline}/drift": {
      "get": {
        "tags": [
          "api"
        ],
        "summary": "Detected schema changes vs contract",
        "operationId": "api_schema_drift_api_schema__pipeline__drift_get",
        "parameters": [
          {
            "name": "pipeline",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pipeline"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DriftEventOut"
                  },
                  "title": "Response Api Schema Drift Api Schema  Pipeline  Drift Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/schema/{pipeline}/accept": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "Accept drift as new contract",
        "operationId": "api_schema_accept_api_schema__pipeline__accept_post",
        "parameters": [
          {
            "name": "pipeline",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pipeline"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Api Schema Accept Api Schema  Pipeline  Accept Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DriftEventOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "pipeline": {
            "type": "string",
            "title": "Pipeline"
          },
          "drift": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Drift"
          },
          "detected_at": {
            "type": "string",
            "title": "Detected At"
          },
          "accepted": {
            "type": "boolean",
            "title": "Accepted"
          }
        },
        "type": "object",
        "required": [
          "id",
          "pipeline",
          "drift",
          "detected_at",
          "accepted"
        ],
        "title": "DriftEventOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}